+++ programs/echo_server_upcall.c
@@ -40,6 +40,7 @@
 #include <string.h>
 #include <stdarg.h>
 #include <sys/types.h>
+#include <io.h>
 #ifndef _WIN32
 #include <unistd.h>
 #include <sys/socket.h>
@@ -141,7 +142,7 @@ handle_upcall(struct socket *sock, void *data, int flgs)
 					snd_info.snd_ppid = rn.recvv_rcvinfo.rcv_ppid;
 					snd_info.snd_context = 0;
 					snd_info.snd_assoc_id = rn.recvv_rcvinfo.rcv_assoc_id;
-					if (usrsctp_sendv(sock, buf, n, NULL, 0, &snd_info, sizeof(struct sctp_sndinfo), SCTP_SENDV_SNDINFO, 0) < 0) {
+					if (usrsctp_sendv(sock, buf, (size_t) n, NULL, 0, &snd_info, sizeof(struct sctp_sndinfo), SCTP_SENDV_SNDINFO, 0) < 0) {
 						perror("sctp_sendv");
 					}
 				}
